+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### todo
 - check whether number input is valid in init
 
+### v0.0.2
+- Added Bubble sort to start working with subject operations
+- Added operations int in t_stacks to keep track of operations count
+
 ### v0.0.1
 - Added README.md with subject
 - Added ft_printf + libft submodules & header file